+++ Makefile
@@ -4,19 +4,19 @@ LDFLAGS= -g -pthread -lrt
 target=moth
 
 WORK_HOME=/home/misery/moth
-FRAME_DIR=${WORK_HOME}/frame
-FRAME_SRC=${FRAME_DIR}/src
-FRAME_INC=${FRAME_DIR}/include
+BASE_DIR=${WORK_HOME}/base
+BASE_SRC=${BASE_DIR}/src
+BASE_INC=${BASE_DIR}/include
 SRCS=$(wildcard *.cpp ${WORK_HOME}/lru_cache/*.cpp \
-		${WORK_HOME}/ftp_client/*.cpp ${FRAME_SRC}/common/*.cpp \
-		${FRAME_SRC}/session/*.cpp ${FRAME_SRC}/message/*.cpp ${FRAME_SRC}/sys/*.cpp ${FRAME_SRC}/db/*.cpp \
-		${FRAME_SRC}/log/*.cpp ${FRAME_SRC}/exception/*.cpp ${FRAME_SRC}/config/*.cpp)
+		${WORK_HOME}/ftp_client/*.cpp ${BASE_SRC}/common/*.cpp \
+		${BASE_SRC}/session/*.cpp ${BASE_SRC}/message/*.cpp ${BASE_SRC}/sys/*.cpp ${BASE_SRC}/db/*.cpp \
+		${BASE_SRC}/log/*.cpp ${BASE_SRC}/exception/*.cpp ${BASE_SRC}/config/*.cpp)
 		
 OBJS=$(patsubst %cpp,%o,$(SRCS))
 
-INCLUDE=-I${WORK_HOME}/lru_cache/ -I${WORK_HOME}/ftp_client/ -I${FRAME_INC}/sys/ \
-		-I${FRAME_INC}/common/ -I${FRAME_INC}/session/ -I${FRAME_INC}/message/ -I${FRAME_INC}/db/ \
-		-I${FRAME_INC}/log/ -I${FRAME_INC}/exception/ -I${FRAME_INC}/config/
+INCLUDE=-I${WORK_HOME}/lru_cache/ -I${WORK_HOME}/ftp_client/ -I${BASE_INC}/sys/ \
+		-I${BASE_INC}/common/ -I${BASE_INC}/session/ -I${BASE_INC}/message/ -I${BASE_INC}/db/ \
+		-I${BASE_INC}/log/ -I${BASE_INC}/exception/ -I${BASE_INC}/config/
 
 all : $(target)
 
@@ -30,6 +30,6 @@ moth : $(OBJS)
 
 clean:
 	@rm -rf $(target) *.o ${WORK_HOME}/lru_cache/*.o ${WORK_HOME}/ftp_client/*.o \
-	${FRAME_SRC}/sys/*.o ${FRAME_SRC}/common/*o ${FRAME_SRC}/session/*.o ${FRAME_SRC}/message/*.o \
-	${FRAME_SRC}/db/*.o ${FRAME_SRC}/log/*.o ${FRAME_SRC}/exception/*.o ${FRAME_SRC}/config/*.o
+	${BASE_SRC}/sys/*.o ${BASE_SRC}/common/*o ${BASE_SRC}/session/*.o ${BASE_SRC}/message/*.o \
+	${BASE_SRC}/db/*.o ${BASE_SRC}/log/*.o ${BASE_SRC}/exception/*.o ${BASE_SRC}/config/*.o
