+++ base/src/common/util.cpp
@@ -1,0 +1,33 @@
+#include "util.h"
+
+UTILS_NS_BEGIN
+
+bool get_work_dir(char* buf)
+{
+	if (NULL == buf)
+	{
+		return false;
+	}
+	
+    char* workdir = NULL;
+    workdir = getenv(WORK_DIR);
+    if (NULL == workdir)
+    {
+        return false;
+    }
+    else
+    {
+        strcpy(buf, workdir);
+        unsigned int len = strlen(buf);
+		// 环境变量的最后不要加'/'
+        if (buf[len - 1] == '/')
+		{
+			buf[len - 1] = 0;
+		}
+		
+        return true;
+    }
+}
+
+UTILS_NS_END
+
