+++ base/src/sys/base.cpp
@@ -1,0 +1,16 @@
+#include "base.h"
+
+SYS_NS_BEGIN
+
+CBase::CBase()
+{
+
+}
+
+CBase::~CBase()
+{
+
+}
+
+SYS_NS_END
+
