+++ base/src/sys/thread_pool.cpp
@@ -1,0 +1,12 @@
+#include "thread_pool.h"
+
+SYS_NS_BEGIN
+
+
+SYS_NS_END
+
+
+// int main()
+// {
+//		return 0;
+// }
