--- frame/src/log/log.cpp
@@ -1,1 +1,0 @@
-#include "log.h"
