--- frame/src/sys/base.cpp
@@ -1,16 +1,0 @@
-#include "base.h"
-
-SYS_NAMESPACE_BEGIN
-
-CBase::CBase()
-{
-
-}
-
-CBase::~CBase()
-{
-
-}
-
-SYS_NAMESPACE_END
-
