--- frame/src/sys/thread_pool.cpp
@@ -1,12 +1,0 @@
-#include "thread_pool.h"
-
-SYS_NAMESPACE_BEGIN
-
-
-SYS_NAMESPACE_END
-
-
-// int main()
-// {
-//		return 0;
-// }
